Adds `--[no-]interactive` option to `bin/install` script
In non-interactive mode (`--no-interactive`) the script will pass extra vars to
Ansible, eliminating pauses for interaction.

Another way of accomplishing that would be to make Ansible believe there is no
TTY, in which case it will use all the defaults defined in the playbook. I'm not
sure how to do that yet though.