Updates Ansible Tower playbook to allow re-running
Previously, many of the Ansible Tower plays would fail while trying to create an
object that already exists. This commit updates many plays to PUT data directly
to the appropriate URL, rather than POST it to the top-level, which essentially
turns a "create" into a "create-or-update."

The net effect is that the Ansible Tower playbook should be idempotent now.